Add and use has_ended helper testing if a Sanction is expired
This helper is mostly used in places where we infer a topic assignment
(is this entity currently sanctioned or debarred?) from the end date.

In a few places, this commit also changes what "now" date is used as a
comparison. In the helper, we use settings.RUN_TIME. This is more
consistent than using datetime.now() (used in a couple places), which
changes throughout the run. It is more semantically correct than using
context.data_time, as our topic assignments should reflect whether an
entity is debarred *now* (for some definition of now) rather than
at the time when the data was published.