[pl_finanse_sanctions] Use provided XLSX list instead of our own copy
I tackled this because the issues paged raised the warning that the page
hash had changed, indicating an update to the page. I thought that they
had addded an XLSX, but it seems it was something else. The warning has
since disappeared, so it seems this was a temporary change to the site.

Anwway, parsing from the XLSX is nicer anyway