[gb_hmt_sanctions] Modernize and fix splitting logic
This fixes a data bug where we were dropping values. If you look hard
enough at split_items, you'll see that the last item in a list never
gets added. Same with values that are not a list. This means that we're
currently dropping quite a few values for title, birthPlace, position,
passportNumber and idNumber.