[helpers] Add DeprecationWarning to h.extract_date
We would like to use h.apply_date in the future. It uses the date format
specified in the dataset YAML metadata and raises more warnings.

Remove it from one more crawler, which is inactive anyway, but by the
time I noticed that, it was already done. No way to test it since the
datasource apparently is gone.

Fixes https://github.com/opensanctions/operations/issues/877