Allow validating the number of property manifestation for a type
In the metadata YAML, an assertion specification like this means that at 10 Person entities have at least one birthPlace property value.

min:
  property_values:
    Person:
      birthPlace: 10

This carries a slight ambiguity with regards to inheritance: Can an
assertion on LegalEntity.country be satisfied by Person and Company
entities? At the moment, the implementation does not allow this.

This adds a new section to statistics.json which looks like this:

	"things": {
	    "total": 698,
	    [...]
	    "property_values": [
	      {
		"schema": "Company",
		"property": "address",
		"count": 207
	      },
	      {
		"schema": "Company",
		"property": "country",
		"count": 199
	      },
	      {
		"schema": "Company",
		"property": "name",
		"count": 202
	      },
	      {
		"schema": "Company",
		"property": "sector",
		"count": 201
	      },
	      {
		"schema": "Company",
		"property": "taxNumber",
		"count": 199
	      },
	      {
		"schema": "Company",
		"property": "topics",
		"count": 199
	      },
	      {
		"schema": "Person",
		"property": "address",
		"count": 508
	      },
	      [...]

Fixes https://github.com/opensanctions/opensanctions/issues/879.