[cl_info_probidad] Ignore one more field
Some declarations now carry a "Periodo" field, which presumably lists
what period of occupancy this declaration applies to, as people
submit multiple of these declarations throughout the occupancy. We
already capture the start date of the occupancy, so presumably this is
to be ignored like we ignore the submission date of the declaration.