[logs] Fix logging to issues
The issue was that format_json would move level -> severity in the event
dict and subsequently let issues_log error out.

The comment in there about any of these processors performing
stringifcation (which made me not touch it previously) is a lie. Non-str
objects in the log dict simply won't be redacted. This commit doesn't
fix that, but doesn't change it either

I feel this thing is up for a bit of a refactor. Some of the stuff in
there (like the database URI redaction) is a bit scary.