[sg_mas_investor_alert] Never emit broken relationships (#2111)
* [sg_mas_investor_alert] Never emit broken relationships

In the previous version, we would still emit them until somebody saw the
warning, came in, added to the list. But we can't really do anything
else, and if they show up later, we'll still skip them. So do this
instead, where we remember what source IDs we've seen, and don't emit
the relation if we didn't see it.