[ru_egrul] Ingest all archives to build historic Ownership/Directoship (#1815)
See https://github.com/opensanctions/opensanctions/issues/847

The general idea is: We parse all the daily zips into a dataframe (which we save locally as a parquet to be able to resume, but this is more of a convenience thing). We take 2022-01-01 as a base and subsequently update it with the companies from the daily updates. (Actually, we first create monthly partial packages, merge these into a yearly, then fold that into the full cause it's faster than doing 365 joins on the full dataset) Updating means just using the new version of the company, but keeping the no-longer-present Ownership and Directorship relations around and setting and end_date on them.

At the end, we write it all out to CSV, which is then ingested by the crawler in zavod.