Sanction key and interval date to distinguish values
We often get the same company multiple times for different infractions
at different times. Sometimes different owners and other related entities
are listed. We want the sanctions distinguished, and a bit more clarity
why different owners are listed, even though we only know the point in time,
not the period of ownership. Also use the Ownership entity, not property.