Close context before writing delta/index/catalog
I think this may fix the issue where issues in the index are 0 even
though the exporter logged some and they're in the issues.json. My guess
is that they're still in the write buffer of Context at the time that
`write_dataset_index` is called

See https://github.com/opensanctions/opensanctions/issues/2459

`eu_fsf` is a good candidate to verify this, currently two issues but
doesn't show up in our overview, only on the actual issues page.